+++ src/bootstrap4/components.py
@@ -6,14 +6,14 @@ from bootstrap4.utils import render_tag
 from .text import text_value
 
 
-def render_alert(content, alert_type=None, dismissable=True):
+def render_alert(content, alert_type=None, dismissible=True):
     """Render a Bootstrap alert."""
     button = ""
     if not alert_type:
         alert_type = "info"
     css_classes = ["alert", "alert-" + text_value(alert_type)]
-    if dismissable:
-        css_classes.append("alert-dismissable")
+    if dismissible:
+        css_classes.append("alert-dismissible")
         close = _("close")
         button = f'<button type="button" class="close" data-dismiss="alert" aria-label="{close}">&times;</button>'
     button_placeholder = "__BUTTON__"