+++ HISTORY.rst
@@ -14,6 +14,7 @@ History
 - Errors on file inputs are shown (#146)
 - Only display non-field errors in form alert block (#161)
 - Reinstate `bootstrap4_title` block to `bootstrap4.html` (#156)
+- Fix typo in `alert-dismissible` class (#142)
 
 0.0.8 (2019-03-12)
 ++++++++++++++++++